Вынес модели SQLAlchemy в один модуль в database
