+++ src/database/orm/sqlalchemy/models.py
@@ -1,5 +1,5 @@
-from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy import ForeignKey
+from sqlalchemy.orm import Mapped, mapped_column
 
 from database.orm.sqlalchemy.stuff import Base
 from database.annotated_types import (
@@ -7,7 +7,19 @@ from database.annotated_types import (
     intpk,
     utcnow
 )
-from .schemas import CVModelEnum
+from roles.schemas import Role
+from cv_models.schemas import CVModelEnum
+
+
+class User(Base):
+
+    __tablename__ = 'users'
+
+    id: Mapped[intpk]
+    name: Mapped[str] = mapped_column(unique=True)
+    role: Mapped[Role]
+    token_amount: Mapped[int]
+    registration_date: Mapped[utcnow]
 
 
 class CVModelTable(Base):
@@ -19,9 +31,9 @@ class CVModelTable(Base):
     cost: Mapped[int]
 
 
-class ImageHistory(Base):
+class TaskHistory(Base):
 
-    __tablename__ = 'image_histories'
+    __tablename__ = 'task_histories'
 
     id: Mapped[intpk]
     user_id: Mapped[int] = mapped_column(ForeignKey('users.id'))
@@ -36,5 +48,13 @@ class Task(Base):
     id: Mapped[intpk]
     cv_model_id: Mapped[int] = mapped_column(ForeignKey('cv_models.id'))
     msg_id: Mapped[int] = mapped_column(unique=True)
-    result_path: Mapped[str]
+    image_id: Mapped[int] = mapped_column(ForeignKey('images.id'))
     result_sum: Mapped[float | None]
+
+
+class Image(Base):
+
+    __tablename__ = 'images'
+
+    id: Mapped[intpk]
+    path: Mapped[str]
